+++ devicemanagement/CHANGELOG.md
@@ -8,10 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Updating the test plan for resetSOC tests.
 ### Changed
 - [SW-14288] Enabled back setAndGetModuleVoltage test.
 ### Deprecated
 ### Removed
+- Removed device-fw-test-plan.ods.
 ### Fixed
 ### Security