+++ devicemanagement/CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Deprecated
 ### Removed
 - Removed device-fw-test-plan.ods.
+- [SW-14210] Fixed bug in setThrottlePowerStatus test
 ### Fixed
 ### Security