+++ devicemanagement/CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added NOPARALLEL label for tests which should not be run in parallel regression.
 ### Changed
 - [SW-14288] Enabled back setAndGetModuleVoltage test.
+- Updated deviceApi version in Conan requirements.
 ### Deprecated
 ### Removed
 - Removed device-fw-test-plan.ods.